Update rect fields to handle signed values.
i3 defines these fields as uint32_t; Sway defines them as (signed) int. In
practice, it's possible to have negative x and y values if floating nodes are
partially offscreen, or with certain combinations of border settings and a
fullscreen window.

i64 should cover all the possibilities here: it's a strict superset of uint32_t
and either a superset of or the same as int, depending on the platform and
architecture.